tests/wpt/tests/css/css-anchor-position/anchor-name-style-contained.html

<!DOCTYPE html>
<title>CSS Anchor Positioning: anchor-name is style contained</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<script src="support/test-common.js"></script>
<style>
  .container {
    position: relative;
  }
  .anchor {
    width: 100px;
    height: 100px;
    anchor-name: var(--anchor-name);
  }
  .contain-style {
    contain: style;
  }
  .target {
    position: absolute;
    position-anchor: var(--anchor-name);
    top: anchor(bottom, 50px);
  }
  #a1 { --anchor-name: --a1; }
  #a2 { --anchor-name: --a2; }
  #a3 { --anchor-name: --a3; }
  #a4 { --anchor-name: --a4; }
  #a5 { --anchor-name: --a5; }
</style>
<body onload="checkLayoutForAnchorPos('.target')">
<div id="a1" class="container">
  <div class="anchor"></div>
  <div class="contain-style">
    <div class="anchor"></div>
  </div>
  <div class="target" data-offset-y="100"></div>
</div>
<div id="a2" class="container">
  <div class="anchor"></div>
  <div class="anchor"></div>
  <div class="contain-style">
    <div class="target" data-offset-y="50"></div>
  </div>
</div>
<div id="a3" class="container">
  <div class="contain-style">
    <div class="anchor"></div>
    <div class="target" data-offset-y="100"></div>
  </div>
</div>
<div id="a4" class="container">
  <div class="anchor contain-style">
    <div class="target" data-offset-y="50"></div>
  </div>
</div>
<div id="a5" class="container">
  <div class="anchor contain-style"></div>
  <div class="target" data-offset-y="100"></div>
</div>
